Port extract method tests to the new framework
