+++ src/VisualStudio/IntegrationTest/New.IntegrationTests/InProcess/EditorVerifierInProcess.cs
@@ -211,6 +211,7 @@ namespace Roslyn.VisualStudio.IntegrationTests.InProcess
                         {
                             FeatureAttribute.Workspace,
                             FeatureAttribute.LightBulb,
+                            FeatureAttribute.Rename,
                         },
                         cancellationToken);
