--- src/VisualStudio/IntegrationTest/TestUtilities/OutOfProcess/InlineRenameDialog_OutOfProc.cs
@@ -1,13 +1,0 @@
-﻿// Licensed to the .NET Foundation under one or more agreements.
-// The .NET Foundation licenses this file to you under the MIT license.
-// See the LICENSE file in the project root for more information.
-
-using Microsoft.CodeAnalysis.Editor.Implementation.InlineRename.HighlightTags;
-
-namespace Microsoft.VisualStudio.IntegrationTest.Utilities.OutOfProcess
-{
-    public static class InlineRenameDialog_OutOfProc
-    {
-        public static string ValidRenameTag => RenameFieldBackgroundAndBorderTag.TagId;
-    }
-}
